CodeMountain: Initial
This is the starting commit of CodeMountain online judge. I have some
plans for this. The core server will be written in Rust(Performance).
Frontend will be written in Typescript with React/Vue as the framework.
There will be a judging engine which will evaluate given problem.There
will be couple of process of judging engine running in the server. The
judge server will take request with the problem id(will talk about that
later) and the actual code to evaluate. The judge server will then
download the testcase data and send them with the code to the RabbitMQ
server. When the judge engine process becomes free, it'll evaluate the
program and send a reply to the message with details about evaluation
like "Is the code 'ACCEPTED'?". The judge server will then get the reply
and send it to the api with a secret token(Only for the judge server and
available through the env). The api will verify the details and if
successful, will update the database. The frontend will poll the server
after submission until there is a verdict.

This application will follow MVP architecture. Which means we are not
going to add all the features from start. We're gonna add them on
demand. At first, basic features like contest, problems, categories,
submissions will be added. Later we may make stuff like community page,
comments widget etc.

I hope I and Shanto will be able to learn so much stuff from this
project. Especially when both of us are noobs in web stuff in Rust.

This shouldn't take that much long? oR iS iT?